+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ---------------------------------
 
+### 1.1.0
+- Added more trail customization configuration options:
+	- Replaced ``Trail Width`` with ``Trail Start Width`` and ``Trail End Width`` to allow more customization of trail.
+	- Added trail transparency configuration items ``Trail Start Alpha`` and ``Trail End Alpha``
+- Added more visibility control configuration options:
+	- Replaced ``Add Visibility Control`` with ``Visbility Setting`` to specify what behavior you wish to control trail visibility.
+	- Added ``Visibility Toggle Action`` for valid actions to show/hide the trail when using an action-based Visbility Setting.
+	- Added ``Visibility Toggle Item`` for selecting your item used to show the trail when using an item-based Visbility Setting.
+
 ### 1.0.0
 - Initial Release